fix hist files in hag root when user skips prompt
If none is set on initialization, hag prompts the user for a purpose. But the
user could always CTRL-C out of the prompt and provide none. If they did, the
history files got saved in the hag root directory, instead of a purpose-specific
directory. The short-term patch is a default "unset" purpose.

In the long term, there is a squishier UI/X decision about why the user might
be using CTRL-C (semantic intent), and what the best/least-surprising reaction
is. This is parallel with some questions about how to support cases like:

- I want a temporary tab to goof around in without long-term persistence and
  without polluting the history of one of my existing tabs/purposes
- I want a private session that doesn't persist at all.