save and restore $PWD per purpose
Not sure how durable this feature will prove to be. I was debugging why my
tabs in Terminal.app stopped retaining their directory consistently on restore.
There's some sort of interaction between how terminal.app launches new tabs,
new windows, and the fact that I clobbered the definition of update_terminal_cwd
in /etc/bashrc_Apple_Terminal (which is responsible for printing an escape
that enables Terminal.app to discover what the current directory is).

In theory, I could just re-enable this, or implement something similar in hag
or in my profile. But, there's yet another moving part here; I'd like to be able
to disable the last-login message, but the macos login command checks for
.hushlogin only in the current directory.

Pushing PWD track/restore into hag solves my own problem, but I think it
also provides a general implementation of something that is obvious adjunct
functionality for hag. Hag would feel incomplete if I had to use it in a term
that didn't get this right.